refactor(gpu): implement sparse binding for demand-paged VRAM allocation
Rollback trigger: If Vulkan sparse bindings cause driver crashes or allocation failures on unsupported hardware.
